Change data type to double

+++ OOPEXERCISES/Circle.java
@@ -1,14 +1,14 @@
 package OOPEXERCISES;
 
 public class Circle extends Shape{
-    float radius;
+    double radius;
 
     Circle(float radius) {
         this.radius = radius;
     }
 
     @Override
-    float calculateArea() {
+    double calculateArea() {
         return 22/7 * radius * radius;
     }
 }
